1.完善推荐模块 2.通过 网关+redis+jwt 完成单点登录 3.统一 各个模块的代码规范 4.工具类装换VO和@Accessors(chain = true) 冲突问题，删除@Accessors解决 5.maven继承导致部分模块启动问题，使用<scope>provided</scope>解决 6.全部模块测试完毕

+++ cloud_interaction/src/main/java/com/lyw/cloudInteraction/dto/ReviewsDto.java
@@ -30,7 +30,7 @@ public class ReviewsDto extends BaseDto {
     private String title;
     
     @ApiModelProperty("总体评分，1-5分")
-    private BigDecimal overallRating;
+    private Integer overallRating;
     
     @ApiModelProperty("多维度评分，例如：{\"teaching\": 4.5, \"content\": 4.0, \"assessment\": 3.5}")
     private String dimensionRatings;
