+++ cloud_interaction/src/main/java/com/lyw/cloudInteraction/vo/ReviewsVo.java
@@ -41,7 +41,7 @@ public class ReviewsVo extends BaseVo {
 
     @ApiModelProperty("总体评分，1-5分")
     @TableField(value = "overall_rating")
-    private BigDecimal overallRating;
+    private Integer overallRating;
 
     @ApiModelProperty("多维度评分，例如：{\"teaching\": 4.5, \"content\": 4.0, \"assessment\": 3.5}")
     @TableField(value = "dimension_ratings")
